App Engine - moz-mktg-prod-001/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
google-cloud-storage==1.3.2
google-cloud-bigquery[pandas]==1.3.0
google-auth-oauthlib==0.2.0